Changing DosingPump.regiment.duration to a float64 rather than a duration. This will allow partial second dosing

+++ jsx/doser/calibration_modal.jsx
@@ -27,7 +27,7 @@ export default class CalibrationModal extends React.Component {
 
   calibrate (duration, speed) {
     const payload = {
-      duration: parseInt(duration),
+      duration: parseFloat(duration),
       speed: parseInt(speed)
     }
